% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cpi_df.R
\name{create_cpi_df}
\alias{create_cpi_df}
\title{Creates dataframe of NZ CPI from StatsNZ csv file}
\usage{
create_cpi_df(cpi)
}
\arguments{
\item{cpi}{path to downloaded CPI csv file from StatsNZ}
}
\value{
a data.frame with column 'Date' giving quarterly dates, and column 'CPI_Index' giving NZ consumer price index.
}
\description{
Using consumer price index data from StatsNZ the function reformats the date column, fills in missing data and selects the 'all group' CPI
(which is an average of all the CPI categories) to return a dataframe of the NZ CPI and matching quarterly dates.
This function can be used in the \code{data} argument in the other functions in this package.
}
\details{
From the Stats NZ website e.g 'https://www.stats.govt.nz/large-datasets/csv-files-for-download/'
locate the NZ CPI file which should be labelled something like 'consumers-price-index-quarter here-index-numbers.csv'
and download the csv file for input.
}
\examples{
create_cpi_df(cpi='consumers-price-index-june-quarter-2022-index-numbers.csv')
create_cpi_df('consumers-price-index-june-quarter-2022-index-numbers.csv')
create_cpi_df('~/Downloads/consumers-price-index-june-quarter-2022-index-numbers.csv')

\dontrun{
create_cpi_df('a_random_csv.csv')
}

}
\keyword{CPI,}
\keyword{NZ,}
\keyword{StatsNZ}
\keyword{consumer}
\keyword{index,}
\keyword{price}
