% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_change_cpi.R
\name{perc_change_cpi}
\alias{perc_change_cpi}
\title{Calculates percentage change of NZ CPI between two dates}
\usage{
perc_change_cpi(
  start = data[1, 1],
  end = data[nrow(data), 1],
  data = nz_cpi_data
)
}
\arguments{
\item{start, end}{only accepts a single value chr or date type of format'YYYY-MM-DD', \code{start} (defaults to first date in \code{data}) and \code{end} (defaults to last date in \code{data}) gives
the start and end dates for the period over which you want to calculate how the CPI has changed}

\item{data}{NZ CPI dataset, use \code{create_cpi_df} function or defaults to \code{nz_cpi_data} dataset}
}
\value{
string describing CPI change
}
\description{
Calculates percentage change of NZ CPI between two given dates using the formula (b-a)/a*100 and
returns sentence describing how CPI has changed.
If inputted start and end dates are not listed in the 'Date' column of \code{data} then the function picks
the nearest date in 'Date' to use.
Gives warnings/errors if inputted dates are inappropriate.
}
\examples{
perc_change_cpi()
perc_change_cpi('2000-01-01','1999-01-01')
perc_change_cpi('1999-01-01','2000-01-01')
perc_change_cpi(start='2000-06-01')
perc_change_cpi(start='1900-01-12',end='2020-08-10')
perc_change_cpi(start='1950-01-01',end='2020-09-01', data = create_cpi_df('consumers-price-index-june-quarter-2022-index-numbers.csv'))

\dontrun{
perc_change_cpi('01-01-2000','01-10-2000')
perc_change_cpi(start=c('2000-01-01', '2000-03-01'),end=c('2000-03-01','2000-06-01'))
}

}
\keyword{CPI,}
\keyword{NZ}
\keyword{change,}
\keyword{consumer}
\keyword{index,}
\keyword{percentage,}
\keyword{price}
