% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpi.R
\name{plot_cpi}
\alias{plot_cpi}
\title{Plots NZ CPI}
\usage{
plot_cpi(start = data[1, 1], end = data[nrow(data), 1], data = nz_cpi_data)
}
\arguments{
\item{start, end}{takes chr or date type of format'YYYY-MM-DD', \code{start} (defaults to first date in \code{data}) and \code{end} (defaults to last date in \code{data}) gives the date range of the CPI to plot}

\item{data}{NZ CPI dataset, use \code{create_cpi_df} function or defaults to \code{nz_cpi_data} dataset}
}
\description{
Plots NZ CPI against time.
Optional arguments to the function are start and end dates that limit the displayed CPI to within the date range.
If inputted start and end dates are not listed in the 'Date' column of \code{data} then the function picks the nearest date in 'Date' to use.
Gives warnings/errors if given dates are inappropriate.
}
\details{
Requires ggplot2 and scales.
}
\examples{
plot_cpi()
plot_cpi(data = create_cpi_df('consumers-price-index-june-quarter-2022-index-numbers.csv'))
plot_cpi('1997-01-01','2000-06-01')
plot_cpi(start='1900-01-12',end='2020-08-10')
plot_cpi(end='2000-06-01')
plot_cpi(start='1950-01-01',end='2020-09-01', data = create_cpi_df('consumers-price-index-june-quarter-2022-index-numbers.csv'))

\dontrun{
plot_cpi('01-01-2000','02-10-2000')
plot_cpi(start='2000-01-01',end='1999-01-01')
plot_cpi(start=c('2000-01-01', '2000-03-01'),end=c('1999-01-01','2000-01-01'))
}

}
\keyword{CPI,}
\keyword{NZ}
\keyword{consumer}
\keyword{index,}
\keyword{plot,}
\keyword{price}
